backend/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9
gunicorn==21.2.0